backend/requirements.txt
# ================================
# PERP Backend — Python Dependencies
# ================================

# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0
greenlet==3.1.1

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18

# Validation
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

# Caching & Queue
redis==5.2.1
celery==5.4.0
kombu==5.4.2

# HTTP
httpx==0.28.1

# Utilities
python-dotenv==1.0.1
orjson==3.10.12
pytz==2024.2

# Reporting / Export
reportlab==4.2.5
openpyxl==3.1.5

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
factory-boy==3.3.1
